fix(cat): keep concordance mapper out of server cat module
Move mapCatConcordanceForAiRecommendation into a client-safe module so
the CAT workspace does not import cat.ts and pull Drizzle/pg into the
browser bundle during production builds.
